feat-39: reuse dialog for AI prompting
